refactor: pull GITHUB_OUTPUT into shared script
