+++ scripts/check-changelog.ts
@@ -1,15 +1,7 @@
 import fs from "fs";
+import saveGithubOutput from "./shared/save-github-output";
 
 let result = "SHOULD_NOT_DEPLOY";
-const GITHUB_OUTPUT_LOCATION = process.env.GITHUB_OUTPUT;
-
-if (!GITHUB_OUTPUT_LOCATION) {
-  throw new Error(
-    "Could not find the GITHUB_OUTPUT file. Not in the Github Actions Environment. Abort!"
-  );
-}
-
-const GITHUB_OUTPUT = fs.readFileSync(GITHUB_OUTPUT_LOCATION, "utf8");
 
 const changelog = JSON.parse(
   fs.readFileSync("./frontend/static/changelog.json", "utf8")
@@ -23,8 +15,4 @@ if (
   result = "COMBO_DATA_CHANGED";
 }
 
-fs.writeFileSync(
-  GITHUB_OUTPUT_LOCATION,
-  `${GITHUB_OUTPUT}
-result=${result}`
-);
+saveGithubOutput("result", result);
